refactor: Implement reusable keyword list component
