+++ src/components/signup/KeywordGroup.tsx
@@ -2,10 +2,9 @@ import { colors } from '@/assets/styles/theme';
 import React from 'react';
 import { styled } from 'styled-components';
 import Txt from '../common/text';
-import Keyword from '../common/keyword';
 import { INTEREST_KEYWORD } from '@/utils/constant';
 import { SignUpInfo } from '@/pages/signup/SignUpPage';
-import { getKeywordGroupTitle } from '@/utils/util';
+import KeywordList from '../common/keyword/KeywordList';
 
 type KeywordGroupProps = {
     signUpInfo: SignUpInfo;
@@ -16,7 +15,6 @@ type KeywordGroupProps = {
 export type KeywordType = (keyof typeof INTEREST_KEYWORD)
 
 export default function KeywordGroup({ signUpInfo, setSignUpInfo }: KeywordGroupProps) {
-    const KEYWORD_LIST = Object.keys(INTEREST_KEYWORD) as KeywordType[];
 
     const handleCheckboxChange = (
         e: React.ChangeEvent<HTMLInputElement>,
@@ -58,6 +56,10 @@ export default function KeywordGroup({ signUpInfo, setSignUpInfo }: KeywordGroup
         return totalLength < 10;
     };
 
+    const handleCheck = (keywordGroup: KeywordType, keyword: string) => {
+        return signUpInfo.keywordGroups[keywordGroup].includes(keyword);
+    }
+
     return (
         <KeywordGroupContainer>
             <Txt variant="h5">관심 키워드*</Txt>
@@ -68,27 +70,7 @@ export default function KeywordGroup({ signUpInfo, setSignUpInfo }: KeywordGroup
             >
                 관심 키워드를 최대 10개까지 선택해주세요
             </Txt>
-            <KeywordListContainer>
-                {KEYWORD_LIST.map((keywordGroup) => (
-                    <KeywordGroupBox>
-                        <Txt variant="caption1">{getKeywordGroupTitle(keywordGroup)}</Txt>
-                        <KeywordList>
-                            {INTEREST_KEYWORD[keywordGroup].map((keyword) => {
-                                const ischecked =
-                                    signUpInfo.keywordGroups[keywordGroup].includes(keyword);
-                                return (
-                                    <Keyword
-                                        name={keyword}
-                                        keywordGroup={keywordGroup}
-                                        ischecked={ischecked}
-                                        onChange={handleCheckboxChange}
-                                    />
-                                );
-                            })}
-                        </KeywordList>
-                    </KeywordGroupBox>
-                ))}
-            </KeywordListContainer>
+            <KeywordList handleCheck={handleCheck} handleChange={handleCheckboxChange} />
         </KeywordGroupContainer>
     );
 }
@@ -100,29 +82,4 @@ export const KeywordGroupContainer = styled.div`
     display: flex;
     flex-direction: column;
     position: relative;
-`;
-
-export const KeywordGroupBox = styled.div`
-    width: 100%;
-    height: auto;
-    display: flex;
-    flex-direction: column;
-    gap: 7px;
-    margin-top: 24px;
-`;
-
-export const KeywordListContainer = styled.div`
-    width: 100%;
-    height: auto;
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 42px;
-`;
-
-export const KeywordList = styled.div`
-    width: 100%;
-    height: auto;
-    display: flex;
-    flex-wrap: wrap;
-    gap: 8px;
 `;
